add logoimg to start stage

+++ src/main/java/com/kursova/kep/control/main/MainController.java
@@ -1,23 +1,20 @@
 package com.kursova.kep.control.main;
 
+import com.kursova.kep.control.base.BaseController;
 import com.kursova.kep.custom.CustomProperties;
+import com.kursova.kep.custom.stage.StageUtil;
 import com.kursova.kep.custom.table.TableColumnsGenerator;
 import com.kursova.kep.entity.*;
 import com.kursova.kep.rest.Client;
+import com.kursova.kep.stage.UserTaskStage;
 import javafx.fxml.Initializable;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.TableView;
-import javafx.scene.control.TextArea;
+import javafx.scene.Cursor;
+import javafx.scene.control.*;
 import javafx.scene.input.MouseButton;
 import javafx.scene.input.MouseEvent;
-import javafx.scene.layout.VBox;
+import javafx.stage.Stage;
 
-import java.lang.reflect.Array;
 import java.net.URL;
-import java.sql.Date;
-import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.List;
 import java.util.ResourceBundle;
 
@@ -25,8 +22,9 @@ import java.util.ResourceBundle;
  * Created by Mr. Skip.
  */
 
-public class MainController implements Initializable{
+public class MainController extends BaseController implements Initializable{
     public TableView<Object> table;
+
     private static Class currentClassObject;
     private static BaseEntity addedEntity;
 
@@ -43,30 +41,33 @@ public class MainController implements Initializable{
     public TextArea textArea;
 
     public Label labCurrentTable;
-    public VBox vBox;
-    public List<String> labelStrings;
+    public Label labelUserTask;
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
         currentClassObject = Department.class;
+        Client.setTextArea(textArea);
 
         List<Department> list = Client.get("department", Department[].class).build();
         TableColumnsGenerator.setTableView(table).generateColumns(list, Department.class);
-        Client.setTextArea(textArea);
 
         buttonsHandler();
-        taskLabelsHandler();
+
+        labelUserTask.setCursor(Cursor.HAND);
     }
 
-    private void taskLabelsHandler(){
-        labelStrings = new ArrayList<>();
-        
+    public void setStage(Stage stage) {
+        this.stage = stage;
+        StageUtil.createStageUtil(-1, 22,
+                UserTaskStage.getUserTaskStage(labCurrentTable, table), getStage())
+                .setObject(labelUserTask);
     }
 
     private void buttonsHandler() {
         buttonDelete.addEventFilter(MouseEvent.MOUSE_CLICKED, event -> {
             if (event.getButton() == MouseButton.PRIMARY
-                    && table.getSelectionModel().getSelectedItem() != null){
+                    && table.getSelectionModel().getSelectedItem() != null
+                    && !labCurrentTable.getText().equals("Завдання")){
                 BaseEntity entity = (BaseEntity) table.getSelectionModel().getSelectedItem();
                 Client.delete(currentClassObject.getSimpleName().toLowerCase() + "/" + entity.getId(), currentClassObject);
                 table.getItems().remove(table.getSelectionModel().getSelectedItem());
@@ -74,32 +75,31 @@ public class MainController implements Initializable{
         });
 
         buttonAdd.addEventFilter(MouseEvent.MOUSE_CLICKED, event -> {
-            if (event.getButton() == MouseButton.PRIMARY){
-                try {
-                    if (buttonAdd.getText().equals("Додати")) {
-                        buttonAdd.setText("Зберегти зміни");
-                        buttonAdd.setPrefWidth(200);
-
-                        addedEntity = (BaseEntity) Class.forName(currentClassObject.getPackage().getName() + "." + currentClassObject.getSimpleName())
-                                .newInstance();
-                        addedEntity.setId((long) table.getItems().size() + 1);
-                        table.getItems().add(addedEntity);
-                        table.getSelectionModel().select((long) table.getItems().size() - 1);
-                    }
-                    else {
-                        buttonAdd.setText("Додати");
-                        buttonAdd.setPrefWidth(120);
-                    }
-                } catch (ClassNotFoundException | IllegalAccessException | InstantiationException e) {
-                    e.printStackTrace();
-                }
-
+            if (event.getButton() != MouseButton.PRIMARY
+                    || labCurrentTable.getText().equals("Завдання"))
+                return;
+            try {
                 if (buttonAdd.getText().equals("Зберегти зміни")) {
                     Client.post(currentClassObject.getSimpleName().toLowerCase(), currentClassObject)
                             .setRequest(table.getItems().get(table.getItems().size() - 1))
                             .build();
                     addedEntity = null;
+
+                    buttonAdd.setText("Додати");
+                    buttonAdd.setPrefWidth(120);
                 }
+                else if (buttonAdd.getText().equals("Додати")) {
+                    buttonAdd.setText("Зберегти зміни");
+                    buttonAdd.setPrefWidth(200);
+
+                    addedEntity = (BaseEntity) Class.forName(currentClassObject.getPackage().getName() + "." + currentClassObject.getSimpleName())
+                            .newInstance();
+                    addedEntity.setId((long)table.getItems().size() + 1);
+                    table.getItems().add(addedEntity);
+                    table.getSelectionModel().select((long) table.getItems().size() - 1);
+                }
+            } catch (ClassNotFoundException | IllegalAccessException | InstantiationException e) {
+                e.printStackTrace();
             }
         });
 
@@ -118,12 +118,6 @@ public class MainController implements Initializable{
             String objectName = ((Button) e.getSource()).getText();
 
             currentClassObject = CustomProperties.getClassFromHeaderButtons(objectName);
-
-            if (!table.getColumns().isEmpty()) {
-                table.getColumns().clear();
-                if (!table.getItems().isEmpty())
-                    table.getItems().clear();
-            }
             labCurrentTable.setText(objectName);
             List list = null;
             try {
